Add files via upload
Sentiment Analysis is a strategy extensively used in text mining. Twitter Sentiment Analysis, in this way infers, using advanced message mining techniques to analyze the assessment of the message (here, tweet) as certain, negative and unprejudiced. 
It is generally called Opinion Mining, is on a very basic level for taking apart conversations, ends, and sharing of points of view (all as tweets) for picking business strategy, political examination, and besides for studying public actions. 
Twitter Sentiment Analysis also is used for monitoring and analyzing social phenomena, for predicting potentially dangerous situations and determining the general mood of the blogosphere.